Split `tests/client/rpcNodes`
