+++ tests/client/service/nodesFind.test.ts
@@ -3,20 +3,21 @@ import fs from 'fs';
 import path from 'path';
 import os from 'os';
 import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
-import { Metadata } from '@grpc/grpc-js';
 import { DB } from '@matrixai/db';
-import { KeyManager, utils as keysUtils } from '@/keys';
-import { GRPCServer } from '@/grpc';
-import { NodeConnectionManager, NodeGraph } from '@/nodes';
-import { Sigchain } from '@/sigchain';
-import { ForwardProxy, ReverseProxy } from '@/network';
-import {
-  GRPCClientClient,
-  ClientServiceService,
-  utils as clientUtils,
-} from '@/client';
+import { Metadata } from '@grpc/grpc-js';
+import KeyManager from '@/keys/KeyManager';
+import NodeConnectionManager from '@/nodes/NodeConnectionManager';
+import NodeGraph from '@/nodes/NodeGraph';
+import Sigchain from '@/sigchain/Sigchain';
+import ForwardProxy from '@/network/ForwardProxy';
+import ReverseProxy from '@/network/ReverseProxy';
+import GRPCServer from '@/grpc/GRPCServer';
+import GRPCClientClient from '@/client/GRPCClientClient';
 import nodesFind from '@/client/service/nodesFind';
+import { ClientServiceService } from '@/proto/js/polykey/v1/client_service_grpc_pb';
 import * as nodesPB from '@/proto/js/polykey/v1/nodes/nodes_pb';
+import * as clientUtils from '@/client/utils/utils';
+import * as keysUtils from '@/keys/utils';
 import * as validationErrors from '@/validation/errors';
 import * as testUtils from '../../utils';
 
@@ -130,7 +131,7 @@ describe('nodesFind', () => {
     grpcClient = await GRPCClientClient.createGRPCClientClient({
       nodeId: keyManager.getNodeId(),
       host: '127.0.0.1' as Host,
-      port: grpcServer.port,
+      port: grpcServer.getPort(),
       logger,
     });
   });
