+++ tests/client/service/nodesPing.test.ts
@@ -3,21 +3,23 @@ import fs from 'fs';
 import path from 'path';
 import os from 'os';
 import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
-import { Metadata } from '@grpc/grpc-js';
 import { DB } from '@matrixai/db';
-import { KeyManager, utils as keysUtils } from '@/keys';
-import { GRPCServer } from '@/grpc';
-import { NodeConnectionManager, NodeGraph, NodeManager } from '@/nodes';
-import { Sigchain } from '@/sigchain';
-import { ForwardProxy, ReverseProxy } from '@/network';
-import {
-  GRPCClientClient,
-  ClientServiceService,
-  utils as clientUtils,
-} from '@/client';
+import { Metadata } from '@grpc/grpc-js';
+import KeyManager from '@/keys/KeyManager';
+import NodeConnectionManager from '@/nodes/NodeConnectionManager';
+import NodeGraph from '@/nodes/NodeGraph';
+import NodeManager from '@/nodes/NodeManager';
+import Sigchain from '@/sigchain/Sigchain';
+import ForwardProxy from '@/network/ForwardProxy';
+import ReverseProxy from '@/network/ReverseProxy';
+import GRPCServer from '@/grpc/GRPCServer';
+import GRPCClientClient from '@/client/GRPCClientClient';
 import nodesPing from '@/client/service/nodesPing';
-import * as nodesPB from '@/proto/js/polykey/v1/nodes/nodes_pb';
+import { ClientServiceService } from '@/proto/js/polykey/v1/client_service_grpc_pb';
 import * as utilsPB from '@/proto/js/polykey/v1/utils/utils_pb';
+import * as nodesPB from '@/proto/js/polykey/v1/nodes/nodes_pb';
+import * as clientUtils from '@/client/utils/utils';
+import * as keysUtils from '@/keys/utils';
 import * as validationErrors from '@/validation/errors';
 import * as testUtils from '../../utils';
 
@@ -138,7 +140,7 @@ describe('nodesPing', () => {
     grpcClient = await GRPCClientClient.createGRPCClientClient({
       nodeId: keyManager.getNodeId(),
       host: '127.0.0.1' as Host,
-      port: grpcServer.port,
+      port: grpcServer.getPort(),
       logger,
     });
   });
